code/requirements.txt
multiprocess
matplotlib
pandas
cassandra-driver
argparse
uuid
paho-mqtt
